+++ src/main/java/com/example/demo/repository/review/ReviewRepository.java
@@ -8,14 +8,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.example.demo.domain.review.ReviewImpl;
+import com.example.demo.domain.review.Review;
 
-public interface ReviewRepository extends JpaRepository<ReviewImpl, Long> {
+public interface ReviewRepository extends JpaRepository<Review, Long> {
 
-	@Query("select r from ReviewImpl r join fetch r.member where r.cafe.id = :cafeId")
-	List<ReviewImpl> findAllByCafeId(@Param("cafeId") Long cafeId);
+	@Query("select r from Review r join fetch r.member where r.cafe.id = :cafeId")
+	List<Review> findAllByCafeId(@Param("cafeId") Long cafeId);
 
-	@Query(value = "select r from ReviewImpl r join fetch r.member where r.cafe.id = :cafeId",
-		countQuery = "select count(*) from ReviewImpl r where r.cafe.id = :cafeId")
-	Page<ReviewImpl> findAllWithPagingByCafeId(@Param("cafeId") Long cafeId, Pageable pageable);
+	@Query(value = "select r from Review r join fetch r.member where r.cafe.id = :cafeId",
+		countQuery = "select count(*) from Review r where r.cafe.id = :cafeId")
+	Page<Review> findAllWithPagingByCafeId(@Param("cafeId") Long cafeId, Pageable pageable);
 }
