+++ src/main/java/com/example/demo/service/review/ReviewServiceImpl.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.example.demo.domain.cafe.Cafe;
 import com.example.demo.domain.member.MemberImpl;
-import com.example.demo.domain.review.ReviewImpl;
+import com.example.demo.domain.review.Review;
 import com.example.demo.dto.review.ReviewSaveRequest;
 import com.example.demo.dto.review.ReviewUpdateRequest;
 import com.example.demo.exception.CafegoryException;
@@ -29,19 +29,19 @@ public class ReviewServiceImpl implements ReviewService {
 
 	@Override
 	public Long saveReview(Long memberId, Long cafeId, ReviewSaveRequest request) {
-		ReviewImpl review = ReviewImpl.builder()
+		Review review = Review.builder()
 			.content(request.getContent())
 			.rate(request.getRate())
 			.cafe(findCafeById(cafeId))
 			.member(findMemberById(memberId))
 			.build();
-		ReviewImpl savedReview = reviewRepository.save(review);
+		Review savedReview = reviewRepository.save(review);
 		return savedReview.getId();
 	}
 
 	@Override
 	public void updateReview(Long memberId, Long reviewId, ReviewUpdateRequest request) {
-		ReviewImpl findReview = findReviewById(reviewId);
+		Review findReview = findReviewById(reviewId);
 		MemberImpl findMember = findMemberById(memberId);
 		validateReviewer(findReview, findMember);
 
@@ -49,20 +49,20 @@ public class ReviewServiceImpl implements ReviewService {
 		findReview.updateRate(request.getRate());
 	}
 
-	private static void validateReviewer(ReviewImpl findReview, MemberImpl findMember) {
+	private static void validateReviewer(Review findReview, MemberImpl findMember) {
 		if (!findReview.isValidMember(findMember)) {
 			throw new CafegoryException(ExceptionType.REVIEW_INVALID_MEMBER);
 		}
 	}
 
-	private ReviewImpl findReviewById(Long reviewId) {
+	private Review findReviewById(Long reviewId) {
 		return reviewRepository.findById(reviewId)
 			.orElseThrow(() -> new CafegoryException(REVIEW_NOT_FOUND));
 	}
 
 	@Override
 	public void deleteReview(Long memberId, Long reviewId) {
-		ReviewImpl findReview = findReviewById(reviewId);
+		Review findReview = findReviewById(reviewId);
 		MemberImpl findMember = findMemberById(memberId);
 		validateReviewer(findReview, findMember);
