+++ src/test/java/com/example/demo/domain/review/ReviewTest.java
@@ -15,7 +15,7 @@ class ReviewTest {
 	@Test
 	@DisplayName("null 검증")
 	void validateNull() {
-		assertThatThrownBy(() -> ReviewImpl.builder().content(null).build())
+		assertThatThrownBy(() -> Review.builder().content(null).build())
 			.isInstanceOf(CafegoryException.class);
 	}
 
@@ -23,7 +23,7 @@ class ReviewTest {
 	@ValueSource(strings = {"", " "})
 	@DisplayName("빈값, 공백문자 검증")
 	void validateEmptyOrWhitespace(String value) {
-		assertThatThrownBy(() -> ReviewImpl.builder().content(value).build())
+		assertThatThrownBy(() -> Review.builder().content(value).build())
 			.isInstanceOf(CafegoryException.class);
 	}
 
@@ -31,21 +31,21 @@ class ReviewTest {
 	@ValueSource(doubles = {0, 5})
 	@DisplayName("평점 정상 범위 검증")
 	void validateRateRange(double number) {
-		assertDoesNotThrow(() -> ReviewImpl.builder().content("a").rate(number).build());
+		assertDoesNotThrow(() -> Review.builder().content("a").rate(number).build());
 	}
 
 	@ParameterizedTest
 	@ValueSource(doubles = {-0.1, 5.1})
 	@DisplayName("평점 예외 범위 검증")
 	void validateRateRange_exception(double number) {
-		assertThatThrownBy(() -> ReviewImpl.builder().content("a").rate(number).build())
+		assertThatThrownBy(() -> Review.builder().content("a").rate(number).build())
 			.isInstanceOf(CafegoryException.class);
 	}
 
 	@Test
 	@DisplayName("리뷰 업데이트시, 리뷰 최대 글자수 검증")
 	void validate_update_content_size() {
-		ReviewImpl review = ReviewImpl.builder()
+		Review review = Review.builder()
 			.content("리뷰")
 			.rate(5)
 			.build();
@@ -58,7 +58,7 @@ class ReviewTest {
 	@Test
 	@DisplayName("리뷰 업데이트시, 리뷰 최대 글자수 예외 검증")
 	void validate_update_content_size_exception() {
-		ReviewImpl review = ReviewImpl.builder()
+		Review review = Review.builder()
 			.content("리뷰")
 			.rate(5)
 			.build();
