+++ src/test/java/com/example/demo/helper/ReviewPersistHelper.java
@@ -6,15 +6,15 @@ import javax.persistence.PersistenceContext;
 import com.example.demo.builder.TestReviewBuilder;
 import com.example.demo.domain.cafe.Cafe;
 import com.example.demo.domain.member.MemberImpl;
-import com.example.demo.domain.review.ReviewImpl;
+import com.example.demo.domain.review.Review;
 
 public class ReviewPersistHelper {
 
 	@PersistenceContext
 	private EntityManager em;
 
-	public ReviewImpl persistDefaultReview(Cafe cafe, MemberImpl member) {
-		ReviewImpl review = new TestReviewBuilder().cafe(cafe).member(member).build();
+	public Review persistDefaultReview(Cafe cafe, MemberImpl member) {
+		Review review = new TestReviewBuilder().cafe(cafe).member(member).build();
 		em.persist(review);
 		return review;
 	}
